1) moved to create-react-app format, 2) fixed the toolbar hiding bug
