+++ .gitignore
@@ -1,4 +1,23 @@
-node_modules/
+# See https://help.github.com/ignore-files/ for more about ignoring files.
+
+# dependencies
+/node_modules
+
+# testing
+/coverage
+
+# production
+/build
+
+# misc
+.DS_Store
+.env
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
+
+
+# from original project
 public/index.js
 public/style.css
 config.json
