+++ package.json
@@ -1,44 +1,29 @@
 {
   "name": "IrisMeet",
   "version": "1.0.0",
-  "description": "IrisMeet",
-  "main": "server/app.js",
-  "scripts": {
-    "test": "npm test"
-  },
+  "private": true,
   "author": "Rob Sliwa",
   "license": "MIT",
+  "main": "server/app.js",
   "dependencies": {
-    "compression": "^1.6.2",
-    "express": "^4.14.0",
     "flux": "^2.1.1",
     "font-awesome": "^4.3.0",
     "iris-auth-js-sdk": "1.0.3",
     "iris-react-webrtc": "1.0.19",
     "keymirror": "^0.1.1",
     "lodash": "^3.6.0",
-    "react": "^15.3.0",
-    "react-dom": "^15.3.0",
+    "react": "^15.5.3",
+    "react-dom": "^15.5.3",
     "react-router": "2.6.1",
     "whatwg-fetch": "^2.0.3"
   },
   "devDependencies": {
-    "aliasify": "^1.7.2",
-    "babel-cli": "^6.8.0",
-    "babel-core": "^6.1.2",
-    "babel-polyfill": "^6.0.16",
-    "babel-preset-es2015": "^6.6.0",
-    "babel-preset-react": "^6.1.2",
-    "babel-preset-stage-2": "^6.5.0",
-    "babelify": "^7.2.0",
-    "browserify": "^12.0.1",
-    "gulp": "^3.8.11",
-    "gulp-rename": "^1.2.2",
-    "gulp-server-livereload": "1.3.0",
-    "gulp-sass": "^2.0.1",
-    "gulp-concat": "^2.5.2",
-    "vinyl-source-stream": "^1.1.0",
-    "watchify": "^2.4.0",
-    "node-notifier": "^4.2.1"
+    "react-scripts": "0.9.5"
+  },
+  "scripts": {
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test --env=jsdom",
+    "eject": "react-scripts eject"
   }
 }
