+++ src/components/horizontal-box/horizontal-box.css
@@ -3,7 +3,7 @@
   margin-left: 5px;
   margin-right: 5px;
   display: inline-block;
-  background-color: transparent; //wheat;
+  background-color: transparent; /*wheat;*/
   position: relative;
   width: auto;
   //max-height: 120px;
