+++ src/components/horizontal-wrapper/main-footer.css
@@ -4,5 +4,5 @@
     left: 0;
     right: 0;
     height: 150px;
-    background-color: transparent; //rgba(0, 0, 0, 0.5);
+    background-color: transparent;
 }
