+++ src/components/login-panel/login-panel.css
@@ -35,7 +35,7 @@
 }
 
 #login-panel .form-control:focus {
-  box-shadow: 0px 0px 5px #FFFFFF; //rgba(1,1,1,1.65);
+  box-shadow: 0px 0px 5px #FFFFFF; /*rgba(1,1,1,1.65);*/
   border-color: #FFFFFF;
 }
