+++ src/components/main-video/index.js
@@ -1,5 +1,5 @@
 import React from 'react';
-
+import './main-video.css'
 
 
 function MainVideo(props) {
