+++ src/stores/base-store.js
@@ -1,5 +1,6 @@
 import { EventEmitter } from 'events';
-import AppDispatcher from '../dispatcher/app-dispatcher.jsx';
+//import AppDispatcher from '../dispatcher/app-dispatcher.jsx';
+import AppDispatcher from '../dispatcher/app-dispatcher';
 
 export default class BaseStore extends EventEmitter {
   constructor() {
