+++ src/utils/verify-resolution.js
